requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
azure-functions==1.5.0
azure-functions-durable==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
furl==2.1.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
orderedmultidict==1.0.1
pycparser==2.20
pylint==2.8.2
pyodbc==4.0.30
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wrapt==1.12.1
yarl==1.6.3
